+++ skore/tests/unit/conftest.py
@@ -1,12 +1,14 @@
 import numpy as np
 import pytest
 from sklearn.datasets import make_classification, make_regression
+from sklearn.dummy import DummyClassifier, DummyRegressor
 from sklearn.ensemble import RandomForestClassifier
 from sklearn.linear_model import LinearRegression, LogisticRegression
 from sklearn.model_selection import train_test_split
 from sklearn.pipeline import Pipeline
 from sklearn.preprocessing import StandardScaler
 from sklearn.svm import SVC
+from skore import ComparisonReport, CrossValidationReport, EstimatorReport
 
 
 @pytest.fixture
@@ -271,3 +273,99 @@ def linear_regression_multioutput_with_train_test(
 ):
     X_train, X_test, y_train, y_test = regression_multioutput_train_test_split
     return LinearRegression().fit(X_train, y_train), X_train, X_test, y_train, y_test
+
+
+@pytest.fixture
+def estimator_reports_binary_classification(binary_classification_train_test_split):
+    X_train, X_test, y_train, y_test = binary_classification_train_test_split
+
+    estimator_report_1 = EstimatorReport(
+        DummyClassifier(strategy="uniform", random_state=0),
+        X_train=X_train,
+        y_train=y_train,
+        X_test=X_test,
+        y_test=y_test,
+    )
+    estimator_report_2 = EstimatorReport(
+        DummyClassifier(strategy="uniform", random_state=1),
+        X_train=X_train,
+        y_train=y_train,
+        X_test=X_test,
+        y_test=y_test,
+    )
+
+    return estimator_report_1, estimator_report_2
+
+
+@pytest.fixture
+def comparison_estimator_reports_binary_classification(
+    estimator_reports_binary_classification,
+):
+    estimator_report_1, estimator_report_2 = estimator_reports_binary_classification
+    return ComparisonReport([estimator_report_1, estimator_report_2])
+
+
+@pytest.fixture
+def cross_validation_reports_binary_classification(binary_classification_data):
+    X, y = binary_classification_data
+    cv_report_1 = CrossValidationReport(
+        DummyClassifier(strategy="uniform", random_state=0), X, y
+    )
+    cv_report_2 = CrossValidationReport(
+        DummyClassifier(strategy="uniform", random_state=1), X, y
+    )
+    return cv_report_1, cv_report_2
+
+
+@pytest.fixture
+def comparison_cross_validation_reports_binary_classification(
+    cross_validation_reports_binary_classification,
+):
+    cv_report_1, cv_report_2 = cross_validation_reports_binary_classification
+    return ComparisonReport([cv_report_1, cv_report_2])
+
+
+@pytest.fixture
+def estimator_reports_regression(regression_train_test_split):
+    X_train, X_test, y_train, y_test = regression_train_test_split
+
+    estimator_report_1 = EstimatorReport(
+        DummyRegressor(),
+        X_train=X_train,
+        y_train=y_train,
+        X_test=X_test,
+        y_test=y_test,
+    )
+    estimator_report_2 = EstimatorReport(
+        DummyRegressor(),
+        X_train=X_train,
+        y_train=y_train,
+        X_test=X_test,
+        y_test=y_test,
+    )
+
+    return estimator_report_1, estimator_report_2
+
+
+@pytest.fixture
+def comparison_estimator_reports_regression(
+    estimator_reports_regression,
+):
+    estimator_report_1, estimator_report_2 = estimator_reports_regression
+    return ComparisonReport([estimator_report_1, estimator_report_2])
+
+
+@pytest.fixture
+def cross_validation_reports_regression(regression_data):
+    X, y = regression_data
+    cv_report_1 = CrossValidationReport(DummyRegressor(), X, y)
+    cv_report_2 = CrossValidationReport(DummyRegressor(), X, y)
+    return cv_report_1, cv_report_2
+
+
+@pytest.fixture
+def comparison_cross_validation_reports_regression(
+    cross_validation_reports_regression,
+):
+    cv_report_1, cv_report_2 = cross_validation_reports_regression
+    return ComparisonReport([cv_report_1, cv_report_2])
