+++ skore/tests/unit/displays/metrics_summary/test_comparison_cross_validation.py
@@ -3,83 +3,41 @@
 import pandas as pd
 import pytest
 from pandas.testing import assert_frame_equal, assert_index_equal
-from sklearn.datasets import make_classification, make_regression
-from sklearn.dummy import DummyClassifier, DummyRegressor
+from sklearn.base import clone
+from sklearn.dummy import DummyClassifier
 from sklearn.metrics import accuracy_score, get_scorer
-from skore import ComparisonReport, CrossValidationReport
-from skore._sklearn._plot import MetricsSummaryDisplay
+from skore import ComparisonReport, CrossValidationReport, MetricsSummaryDisplay
 from skore._utils._testing import check_cache_changed, check_cache_unchanged
 
 
-@pytest.fixture
-def classification_data():
-    X, y = make_classification(class_sep=0.1, random_state=42)
-    return X, y
-
-
-def make_classifier():
-    return DummyClassifier(strategy="uniform", random_state=0)
-
-
-@pytest.fixture
-def report(classification_data):
-    """ComparisonReport of CrossValidationReports for classification estimators.
-
-    Note that the two CrossValidationReports do not have the same number of CV splits.
-    """
-    X, y = classification_data
-
-    report = ComparisonReport(
-        [
-            CrossValidationReport(make_classifier(), X, y),
-            CrossValidationReport(make_classifier(), X, y, cv_splitter=3),
-        ]
-    )
-
-    return report
-
-
-@pytest.fixture
-def report_regression():
-    """ComparisonReport of CrossValidationReports for regression estimators."""
-    X, y = make_regression(random_state=42)
-
-    report = ComparisonReport(
-        [
-            CrossValidationReport(DummyRegressor(), X, y),
-            CrossValidationReport(DummyRegressor(), X, y, cv_splitter=3),
-        ]
-    )
-
-    return report
-
-
-def test_aggregate_none(report):
+def test_aggregate_none(comparison_cross_validation_reports_binary_classification):
     """`summarize` works as intended with `aggregate=None`."""
+    report = comparison_cross_validation_reports_binary_classification
     result = report.metrics.summarize(aggregate=None)
     assert isinstance(result, MetricsSummaryDisplay)
     result_df = result.frame()
 
     assert_index_equal(result_df.columns, pd.Index(["Value"]))
     assert result_df.index.names == ["Metric", "Label / Average", "Estimator", "Split"]
-    assert len(result_df) == 64
+    assert len(result_df) == 80
 
 
-def test_aggregate_none_flat_index(report):
+def test_aggregate_none_flat_index(
+    comparison_cross_validation_reports_binary_classification,
+):
     """
     `summarize` works as intended with `aggregate=None` and `flat_index=True`.
     """
-    result = report.metrics.summarize(
-        aggregate=None,
-        flat_index=True,
-    ).frame()
+    report = comparison_cross_validation_reports_binary_classification
+    result = report.metrics.summarize(aggregate=None, flat_index=True).frame()
 
     assert_index_equal(result.columns, pd.Index(["Value"]))
-    assert len(result) == 64
+    assert len(result) == 80
 
 
-def test_default(report):
+def test_default(comparison_cross_validation_reports_binary_classification):
     """`summarize` works as intended with its default attributes."""
+    report = comparison_cross_validation_reports_binary_classification
     result = report.metrics.summarize().frame()
 
     assert_index_equal(
@@ -97,12 +55,13 @@ def test_default(report):
     assert len(result) == 8
 
 
-def test_default_regression(report_regression):
+def test_default_regression(comparison_cross_validation_reports_regression):
     """
     `summarize` works as intended with its default attributes for regression
     models.
     """
-    result = report_regression.metrics.summarize().frame()
+    report = comparison_cross_validation_reports_regression
+    result = report.metrics.summarize().frame()
 
     assert_index_equal(
         result.columns,
@@ -122,31 +81,35 @@ def test_default_regression(report_regression):
     )
 
 
-def test_aggregate_sequence_of_one_element(report):
+def test_aggregate_sequence_of_one_element(
+    comparison_cross_validation_reports_binary_classification,
+):
     """Passing a list of one string is the same as passing the string itself."""
+    report = comparison_cross_validation_reports_binary_classification
     assert_frame_equal(
         report.metrics.summarize(aggregate="mean").frame(),
         report.metrics.summarize(aggregate=["mean"]).frame(),
     )
 
 
-def test_aggregate_is_used_in_cache(report):
+def test_aggregate_is_used_in_cache(
+    comparison_cross_validation_reports_binary_classification,
+):
     """`aggregate` should be used when computing the cache key.
 
     In other words, if you call `summarize` twice with different values of
     `aggregate`, you should get a different result.
     """
+    report = comparison_cross_validation_reports_binary_classification
     call1 = report.metrics.summarize(aggregate="mean").frame()
     call2 = report.metrics.summarize(aggregate=("mean", "std")).frame()
     assert list(call1.columns) != list(call2.columns)
 
 
-def test_scoring(report):
+def test_scoring(comparison_cross_validation_reports_binary_classification):
     """`summarize` works as intended with the `scoring` parameter."""
-    result = report.metrics.summarize(
-        scoring=["accuracy"],
-        aggregate=None,
-    ).frame()
+    report = comparison_cross_validation_reports_binary_classification
+    result = report.metrics.summarize(scoring=["accuracy"], aggregate=None).frame()
 
     assert_index_equal(result.columns, pd.Index(["Value"]))
     assert_index_equal(
@@ -161,14 +124,17 @@ def test_scoring(report):
                 ("Accuracy", "DummyClassifier_2", "Split #0"),
                 ("Accuracy", "DummyClassifier_2", "Split #1"),
                 ("Accuracy", "DummyClassifier_2", "Split #2"),
+                ("Accuracy", "DummyClassifier_2", "Split #3"),
+                ("Accuracy", "DummyClassifier_2", "Split #4"),
             ],
             names=("Metric", "Estimator", "Split"),
         ),
     )
 
 
-def test_favorability(report):
+def test_favorability(comparison_cross_validation_reports_binary_classification):
     """`summarize` works as intended with `indicator_favorability=True`."""
+    report = comparison_cross_validation_reports_binary_classification
     result = report.metrics.summarize(indicator_favorability=True).frame()
 
     assert_index_equal(
@@ -187,9 +153,9 @@ def test_favorability(report):
     assert len(result) == 8
 
 
-def test_cache(report):
+def test_cache(comparison_cross_validation_reports_binary_classification):
     """`summarize` results are cached."""
-
+    report = comparison_cross_validation_reports_binary_classification
     with check_cache_changed(report._cache):
         result = report.metrics.summarize().frame()
 
@@ -199,14 +165,15 @@ def test_cache(report):
     assert_frame_equal(result, cached_result)
 
 
-def test_init_with_report_names(classification_data):
+def test_init_with_report_names(forest_binary_classification_data):
     """
     If the estimators are passed as a dict, then the estimator names are the dict keys.
     """
 
-    X, y = classification_data
-    cv_report1 = CrossValidationReport(make_classifier(), X, y)
-    cv_report2 = CrossValidationReport(make_classifier(), X, y)
+    estimator_1, X, y = forest_binary_classification_data
+    estimator_2 = clone(estimator_1)
+    cv_report1 = CrossValidationReport(estimator_1, X, y)
+    cv_report2 = CrossValidationReport(estimator_2, X, y)
 
     comp = ComparisonReport({"r1": cv_report1, "r2": cv_report2})
 
@@ -221,9 +188,13 @@ def test_init_with_report_names(classification_data):
     )
 
 
-def test_X_y(report, classification_data):
+def test_X_y(
+    comparison_cross_validation_reports_binary_classification,
+    binary_classification_data,
+):
     """`summarize` works as intended with `data_source="X_y"`."""
-    X, y = classification_data
+    report = comparison_cross_validation_reports_binary_classification
+    X, y = binary_classification_data
     result = report.metrics.summarize(data_source="X_y", X=X, y=y).frame()
 
     assert_index_equal(
@@ -241,13 +212,13 @@ def test_X_y(report, classification_data):
     assert len(result) == 8
 
 
-def test_cache_poisoning(classification_data):
+def test_cache_poisoning(binary_classification_data):
     """Computing metrics for a ComparisonReport should not influence the
     metrics computation for the internal CVReports.
 
     Non-regression test for https://github.com/probabl-ai/skore/issues/1706
     """
-    X, y = classification_data
+    X, y = binary_classification_data
 
     report_1 = CrossValidationReport(
         DummyClassifier(strategy="uniform", random_state=1), X=X, y=y
@@ -274,10 +245,11 @@ def test_cache_poisoning(classification_data):
     ],
 )
 def test_comparison_report_cv_report_summarize_scoring_single_list_equivalence(
-    report, scoring, scoring_kwargs
+    comparison_cross_validation_reports_binary_classification, scoring, scoring_kwargs
 ):
     """Check that passing a single string, callable, scorer is equivalent to passing a
     list with a single element."""
+    report = comparison_cross_validation_reports_binary_classification
     result_single = report.metrics.summarize(
         scoring=scoring, scoring_kwargs=scoring_kwargs
     ).frame()
